// frontend/src/features/rdbmsConnection/types.ts
/*
 * Copyright 2026 agwlvssainokuni
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export type RdbmsType = 'MYSQL' | 'MARIADB' | 'POSTGRESQL' | 'H2'

export interface ConnectionSummary {
  id: number
  name: string
  rdbmsType: RdbmsType
  host: string
  databaseName: string
}

export interface ConnectionDetail {
  id: number
  name: string
  rdbmsType: RdbmsType
  host: string
  port: number
  databaseName: string
  username: string
  additionalParams: string | null
}

export interface ConnectionConfig {
  name: string
  rdbmsType: RdbmsType
  host: string
  port: number
  databaseName: string
  username: string
  password: string
  additionalParams: string
}

export interface ConnectionTestResult {
  success: boolean
  message: string
}